% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tools.R
\name{img2array}
\alias{img2array}
\title{Extract the raw data from a image}
\usage{
img2array(img)
}
\arguments{
\item{img}{the magick image object}
}
\value{
A array, the dimensions of the array are height, width,channel
}
\description{
Extract the raw data from a image object, return an array object
}
