% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tools.R
\name{img2df}
\alias{img2df}
\title{Transfrom the image to dataframe}
\usage{
img2df(img)
}
\arguments{
\item{img}{the magick image object}
}
\value{
A dataframe, the columns are the channels of the image, row are the pixels
}
\description{
Extract the raw data from a image object, return an data.frame object
}
