% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tools.R
\name{nb_col}
\alias{nb_col}
\title{Automate select the number of colours from an image}
\usage{
nb_col(imgpath)
}
\arguments{
\item{imgpath}{the path of image}
}
\value{
a ggplot2 obejct, containing the boot-variance result.
}
\description{
Using the variance-bootstrap method to automate select the number of colours
}
\seealso{
\code{\link[factoextra:fviz_nbclust]{factoextra::fviz_nbclust()}}
}
