% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_function.R
\name{pal_ipcc_multihue}
\alias{pal_ipcc_multihue}
\title{export}
\usage{
pal_ipcc_multihue(
  palette = c("blue3", "blue4", "blue5", "purple3", "purple4", "purple5", "red3",
    "red4", "red5", "green3", "green4", "green5")
)
}
\description{
export
}
