% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_sort.R
\name{palette_sort}
\alias{palette_sort}
\title{Sort the palette by hand}
\usage{
palette_sort(pal, simple = FALSE)
}
\arguments{
\item{pal}{the palette list object returned by palfromimg}
}
\value{
A list object, contain a ggplot2 palette function and meta info of the image
}
\description{
This shiny APP can handle the order of the palette by hand
}
\examples{
pal_deus <- palfromimg(imgpath = "./ipcc10.jpg", coln =10, sort = 2)
# the order of the palette is not optimized for gradient scale, re-order it
pal_deus_new <- palette_sort(pal_deus)
# order the palette by hand in the App window
}
\seealso{
\code{\link[=palfromimg]{palfromimg()}}
}
