% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_from_image.R
\name{palfromimg}
\alias{palfromimg}
\title{Extract color palette from a image file}
\usage{
palfromimg(
  imgpath,
  coln = 6,
  alpha = 1,
  seed = 1234,
  resize = TRUE,
  sort = 1,
  onlycolour = FALSE
)
}
\arguments{
\item{imgpath}{the path of the image file, only support image with RGB color space now}

\item{coln}{the number of colors extracted from the image}

\item{alpha}{the alpha value the palette}

\item{seed}{random seed of kmeans}

\item{resize}{resize(shrink) the image before kmeans to avoid memory overflow in Kmeans}

\item{sort}{sort the palette by color channel}

\item{onlycolour}{if TRUE, return the colour code instead of colour list object}
}
\value{
A list object/character vector object, contain a ggplot2 palette function and meta info of the image/only colour code
}
\description{
This function use the Kmeans method to extract the color palette from a image file
}
\examples{
pal_deus <- palfromimg(imgpath = "./ipcc10.jpg", coln =10, sort = 2)
show_col(pal_deus$pal(10))
}
