% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tools.R
\name{resize_img}
\alias{resize_img}
\title{Resize(Shrink) the image}
\usage{
resize_img(imgpath, resize = "50", write = F, writepath = "img_resize.jpg")
}
\arguments{
\item{imgpath}{the path of image file}

\item{resize}{resize proportionally to width: default 50 px}

\item{write}{logical, if TRUE, write the image object to file}

\item{writepath}{the path of writed image}
}
\value{
A magick object or image_write function
}
\description{
Resize the image to smaller size
}
