% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_from_image.R
\name{scale_colour_mypal}
\alias{scale_colour_mypal}
\title{Generate a ggplot2 color palette scale function (aes = colour)}
\usage{
scale_colour_mypal(palette, discrete = TRUE, ...)
}
\arguments{
\item{palette}{the palette object returned by palfromimg}

\item{discrete}{the discrete scale or gradient scale}
}
\description{
Generate a color scale function by using the palette extracted from a image
}
\examples{
pal_deus <- palfromimg(imgpath = "./ipcc10.jpg", coln =10, sort = 2)
pal_deus$pal(10) \%>\% show_col()
scale_fill_deus <- scale_fill_mypal(pal_deus)
# two ways
fig_iris <- ggplot(data = iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
geom_boxplot() +
scale_fill_deus
# or
fig_iris <- ggplot(data = iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
geom_boxplot() +
scale_colour_mypal(pal_deus, discrete = FALSE)
}
\seealso{
\code{\link[=palfromimg]{palfromimg()}}, \code{\link[=scale_fill_mypal]{scale_fill_mypal()}}
}
